fix: validate incoming package.json's repository field (#295)
When a bad repository field value is present in the incoming package.json,
Wombat Dressing Room should reject the publication request.

This is done by validating the repository permission in the repository
value written in the incoming package.json, in addition to the existing
validation on the repository value in the latest released version.

b/427764808